Moving the VertexStoreRecovery to only depend on the Database (not the entire StateManager).
