+++ core-rust-bridge/src/main/java/com/radixdlt/recovery/VertexStoreRecovery.java
@@ -65,20 +65,25 @@
 package com.radixdlt.recovery;
 
 import com.google.common.reflect.TypeToken;
+import com.radixdlt.database.Database;
 import com.radixdlt.lang.Option;
 import com.radixdlt.lang.Tuple;
 import com.radixdlt.monitoring.LabelledTimer;
 import com.radixdlt.monitoring.Metrics;
 import com.radixdlt.sbor.Natives;
-import com.radixdlt.statemanager.StateManager;
 import java.util.Optional;
 
 public final class VertexStoreRecovery {
-  public VertexStoreRecovery(Metrics metrics, StateManager stateManager) {
+  public VertexStoreRecovery(Metrics metrics, Database database) {
     LabelledTimer<Metrics.MethodId> timer = metrics.stateManager().nativeCall();
-    this.getVertexStore = // WIP?
-        Natives.builder(stateManager, VertexStoreRecovery::getVertexStore)
+    this.getVertexStore =
+        Natives.builder(database, VertexStoreRecovery::getVertexStore)
             .measure(timer.label(new Metrics.MethodId(VertexStoreRecovery.class, "getVertexStore")))
+            .build(new TypeToken<>() {});
+    this.saveVertexStoreFunc =
+        Natives.builder(database, VertexStoreRecovery::saveVertexStore)
+            .measure(
+                timer.label(new Metrics.MethodId(VertexStoreRecovery.class, "saveVertexStore")))
             .build(new TypeToken<>() {});
   }
 
@@ -86,7 +91,15 @@ public final class VertexStoreRecovery {
     return this.getVertexStore.call(Tuple.tuple()).toOptional();
   }
 
-  private static native byte[] getVertexStore(StateManager stateManager, byte[] payload);
+  private static native byte[] getVertexStore(Database database, byte[] payload);
 
   private final Natives.Call1<Tuple.Tuple0, Option<byte[]>> getVertexStore;
+
+  public void saveVertexStore(byte[] vertexStoreBytes) {
+    this.saveVertexStoreFunc.call(vertexStoreBytes);
+  }
+
+  private static native byte[] saveVertexStore(Database database, byte[] payload);
+
+  private final Natives.Call1<byte[], Tuple.Tuple0> saveVertexStoreFunc;
 }
